requirements.txt
Flask==3.0.2
waitress==2.1.2
python-dotenv==1.0.1
